feat: separate metrics server components
